Render canonical per-member shares in expense confirmation
Previous commit deferred per-member share rendering with the justification
"would diverge from on-ledger amounts under integer rounding edge cases".
That was punting: the divergence is structural only if we re-implement the
split, not if we call the same function the record path uses.

- Expose `compute_expense_owed_amounts` from walicord-application as a public
  wrapper around the existing `distribute_owed_amounts` (the function
  `RecordableExpenseAuthoring::new` already uses to derive `owed_by`). Calling
  the same function from the confirmation render guarantees the previewed
  shares equal the on-ledger amounts by construction.

- Change `ExpenseConfirmationParticipantRow::weight` from u32 to u64 so the
  domain `Weight(u64)` flows through without a presentation cast and the
  legacy `u32::try_from(...).expect(...)` site disappears.

- Add `ExpenseAuthoring` variant + delegating `From` to
  `InternalLedgerRouteError` so authoring failures surface as typed errors
  via `?` instead of being absorbed into a misleading preview.

- Confirmation rendering: build `Vec<MemberWeight>` from the resolved
  selection, call `compute_expense_owed_amounts`, index the result by member,
  and pass the per-member share into `ExpenseConfirmationParticipantRow`.
  Members the canonical distribution drops (weight 0 / rounded to 0 share)
  fall through `share_amount = None`, which the existing row template renders
  as `(取り分なし)` / `(端数で0)` — no fabricated share value.

+++ walicord-presentation/src/discord_ledger/pickers.rs
@@ -114,7 +114,10 @@ impl ExpenseParticipantSourceBadge {
 pub struct ExpenseConfirmationParticipantRow {
     pub display_name: SafeLiteralText,
     pub share_amount: Option<String>,
-    pub weight: u32,
+    /// Per-member weight. Matches the domain `Weight(u64)` so the presentation row
+    /// never has to lossily cast and presentation never silently saturates an
+    /// out-of-range weight from the domain.
+    pub weight: u64,
     pub badges: Vec<ExpenseParticipantSourceBadge>,
     pub defaulted_weight: bool,
 }
